Initial release of ms-hotfix-followup-lihas
This package installs any Windows Updates that may still be left after
the ms-hotfix package (available from UIB for a fee) is run.

Note that there is no dependency hard-coded in the
package so you can run it without the ms-hotfix package, too - but
prepare yourself for a LOOOOONG installation time with many many reboots.
(This package triggers internet downloads of every missing package.
UIB's ms-hotfix package, on the other hand, downloads packages only once
and stores them on the OPSI server, which is clearly the better approach.)